moving search indexer to a webpack compiler
