+++ plugins/search-index.js
@@ -1,22 +1,16 @@
+const SearchIndexWebpackPlugin = require("./search-index-compiler");
+
 module.exports = function indexSearch(nextConfig = {}) {
   const { env, outputLocation, outputName } = nextConfig;
   const { WORDPRESS_HOST } = env;
 
   return Object.assign({}, nextConfig, {
-    webpack(config, options) {
-
-      config.module.rules.push({
-        test: /pages\/index\.js/,
-        use: [
-          {
-            loader: require.resolve('./search-index-loader'),
-            options: {
-              outputLocation,
-              WORDPRESS_HOST
-            }
-          },
-        ],
-      });
+    webpack(config) {
+      config.plugins.push(new SearchIndexWebpackPlugin({
+        host: WORDPRESS_HOST,
+        outputLocation,
+        outputName
+      }));
 
       return config;
     }
