+++ plugins/util.js
@@ -1,0 +1,38 @@
+const fs = require('fs');
+
+/**
+ * promiseToWriteFile
+ */
+
+function promiseToWriteFile(location, content) {
+  return new Promise((resolve, reject) => {
+    fs.writeFile(location, content, (err) => {
+      if ( err ) {
+        reject(err);
+        return;
+      }
+      resolve();
+    })
+  })
+}
+
+module.exports.promiseToWriteFile = promiseToWriteFile;
+
+/**
+ * mkdirp
+ */
+
+function mkdirp(directory) {
+  const split = directory.split('/');
+  let temp = '.';
+
+  split.forEach((dir) => {
+    temp = `${temp}/${dir}`;
+
+    if (!fs.existsSync(temp) ) {
+      fs.mkdirSync(temp);
+    }
+  });
+}
+
+module.exports.mkdirp = mkdirp;
