+++ public/search.json
@@ -1,1 +1,1 @@
-{"generated":1598924940473,"posts":[{"title":"Fry, we have a crate to deliver","slug":"fry-we-have-a-crate-to-deliver","date":"2020-08-24T03:02:31"},{"title":"Hey, you add a one and two zeros to that or we walk!","slug":"hey-you-add-a-one-and-two-zeros-to-that-or-we-walk","date":"2020-08-24T03:02:09"},{"title":"Throw her in the brig","slug":"throw-her-in-the-brig","date":"2020-08-24T03:01:52"},{"title":"I love you, buddy!","slug":"i-love-you-buddy","date":"2020-08-24T03:01:30"},{"title":"You don&#8217;t know how to do any of those","slug":"you-dont-know-how-to-do-any-of-those","date":"2020-08-24T03:01:06"}]}
+{"generated":1598926888737,"posts":[{"title":"Fry, we have a crate to deliver","slug":"fry-we-have-a-crate-to-deliver","date":"2020-08-24T03:02:31"},{"title":"Hey, you add a one and two zeros to that or we walk!","slug":"hey-you-add-a-one-and-two-zeros-to-that-or-we-walk","date":"2020-08-24T03:02:09"},{"title":"Throw her in the brig","slug":"throw-her-in-the-brig","date":"2020-08-24T03:01:52"},{"title":"I love you, buddy!","slug":"i-love-you-buddy","date":"2020-08-24T03:01:30"},{"title":"You don&#8217;t know how to do any of those","slug":"you-dont-know-how-to-do-any-of-those","date":"2020-08-24T03:01:06"}]}
